+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Alterações realizadas na extensão.
 ## [1.2.3] - 30/08/2022
 ### Adicionado
 - Informar que a existência de bloco `Inicio`/`Fim` ou `{chaves}` vazio.
+- Novas funções do HCM para manipulação de JSON (`getJSONDecimal`,`getJSONDateTime`,`addKeyAndStringValueInJSON`, entre outras).
 
 ### Modificado
 - Não gerar erro quando encontra um comentário de linha sem conteúdo (`@@`).